+++ LinkEngine/Entities/Enemy.cs
@@ -1,0 +1,10 @@
+﻿namespace LinkEngine.Entities
+{
+    public class Enemy : Entity
+    {
+        public Enemy (int id, string name, int health, int maxHealth) : base (id, name, health, maxHealth)
+        {
+
+        }
+    }
+}