+++ LinkEngine/Entities/Player.cs
@@ -1,0 +1,10 @@
+﻿namespace LinkEngine.Entities
+{
+    public class Player : Entity
+    {
+        public Player(int id, string name, int health, int maxHealth) : base (id, name, health, maxHealth)
+        {
+
+        }
+    }
+}